Adds explicit checks and validation
now checks to ensure that the Animal object s present before attempting to calculate happiness and hunger.

Also now adds checks for food and stroke amount so taht theu can't fall below zero. The data type of those variables
is now also a decimal so that the range is wider when feeding or stroking